% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{coordinates}
\alias{coordinates}
\title{City coordinates data}
\value{
a \sQuote{data.frame} of the city coordinates data.
}
\description{
City coordinates data
}
\seealso{
Other datasets: 
\code{\link{incidence}},
\code{\link{mobility}},
\code{\link{population0}},
\code{\link{population}},
\code{\link{v_by_g_day}}
}
\concept{datasets}
