% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{incidence}
\alias{incidence}
\title{Incidence data}
\value{
a \sQuote{data.frame} of the cases.
}
\description{
Incidence data
}
\seealso{
Other datasets: 
\code{\link{coordinates}},
\code{\link{mobility}},
\code{\link{population0}},
\code{\link{population}},
\code{\link{v_by_g_day}}
}
\concept{datasets}
