% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/li20.R
\name{li20}
\alias{li20}
\title{Build spatPomp object for li20}
\usage{
li20(
  U = 373,
  dt = 1,
  sharedParNames = c("alpha_be", "Beta_be", "alpha_af", "Beta_af", "mu_be", "Z_be",
    "D_be", "mu_af", "Z_af", "D_af", "theta", "tau", "Td_be", "Td_af", "sigma_SE", "E_0",
    "A_0"),
  unitParNames = NULL,
  for_ibpf = T,
  overdispersion_dynamic = F,
  version = c("li20period3"),
  mob_modify_factor = 0,
  days = 30,
  measurement = "simulation"
)
}
\arguments{
\item{U}{A length-one numeric signifying the number of cities to be represented in the spatPomp object.}

\item{dt}{step size, in days, for the euler approximation.}

\item{sharedParNames}{estimated parameters that are equal for each unit.}

\item{unitParNames}{estimated parameters that are different for each unit.}

\item{for_ibpf}{whether this object is generated for ibpf.}

\item{overdispersion_dynamic}{whether to include overdispersion in dynamic model.}

\item{version}{settings to reproduce results from Li et al (2020) or other selected parameter choices}

\item{mob_modify_factor}{A multiplicative factor, which is used to control the magnitude of adjustment for the mobility data}

\item{days}{number of time points, in days}

\item{measurement}{gives alternative sets of measurement model components}
}
\value{
An object of class \sQuote{spatPomp} representing a \code{U}-dimensional spatially coupled Covid POMP model.
}
\description{
Generate a \sQuote{spatPomp} object for reproducing the results of Li et al.
}
