% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/li20period3.R
\name{li20period3}
\alias{li20period3}
\title{Covid spatPomp generator for Jan 10 - Feb 8}
\usage{
li20period3(
  li20_version = "v5",
  U = 373,
  dt = 1/4,
  sharedOneInterval = c("theta", "tau", "sigma_SE", "E_0", "Iu_0"),
  sharedTwoInterval = c("alpha", "Beta", "mu", "Z", "D", "Td"),
  version = c("MLEperiod3", "li20period1", "li20period2", "li20period3"),
  sharedParNames = c("alpha_be", "Beta_be", "alpha_af", "Beta_af", "mu_be", "Z_be",
    "D_be", "mu_af", "Z_af", "D_af", "theta", "tau", "Td_be", "Td_af", "size_a",
    "size_b", "E_0", "Iu_0"),
  unitParNames = NULL,
  for_ibpf = T,
  mob_modify_factor = 20
)
}
\arguments{
\item{li20_version}{choose which version of the model will be used}

\item{U}{A length-one numeric signifying the number of cities to be represented in the spatPomp object}

\item{dt}{step size, in days, for the euler approximation}

\item{sharedOneInterval}{estimated parameters that are equal for each unit and constant throughout time}

\item{sharedTwoInterval}{estimated parameters that are equal for each unit but change on day 15}

\item{version}{choose which set of parameters will be used for code testing}

\item{sharedParNames}{for li20v1, estimated parameters that are equal for each unit}

\item{unitParNames}{for li20v1, estimated parameters that are different for each unit}

\item{for_ibpf}{whether this object is generated for ibpf}

\item{mob_modify_factor}{A multiplicative factor, which is used to control the magnitude of adjustment for the mobility data}
}
\value{
An object of class \sQuote{spatPomp} representing a \code{U}-dimensional spatially coupled Covid POMP model.
}
\description{
Generate a \sQuote{spatPomp} object for Covid in \code{U} cities in China.
}
