% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/li23.R
\name{li23}
\alias{li23}
\title{Build spatPomp object for li23}
\usage{
li23(
  U = 373,
  dt = 1/4,
  sharedOneInterval = c("theta", "tau", "sigma_SE", "E_0", "Iu_0"),
  sharedTwoInterval = c("alpha", "Beta", "mu", "Z", "D", "Td"),
  version = c("MLEperiod3", "li20period1", "li20period2", "li20period3"),
  mob_modify_factor = 20,
  days = 30,
  for_ibpf = T
)
}
\arguments{
\item{U}{A length-one numeric signifying the number of cities to be represented in the spatPomp object.}

\item{dt}{step size, in days, for the euler approximation.}

\item{sharedOneInterval}{estimated parameters that are equal for each unit and constant throughout time}

\item{sharedTwoInterval}{estimated parameters that are equal for each unit but change on day 15}

\item{version}{choose which set of parameters will be used for code testing}

\item{mob_modify_factor}{A multiplicative factor, which is used to control the magnitude of adjustment for the mobility data}

\item{days}{number of time points, in days}

\item{for_ibpf}{whether this object is generated for ibpf.}
}
\value{
An object of class \sQuote{spatPomp} representing a \code{U}-dimensional spatially coupled Covid POMP model.
}
\description{
Generate a \sQuote{spatPomp} object for Covid in \code{U} cities in China.
}
