% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility.R
\name{mobility}
\alias{mobility}
\title{Mobility data}
\value{
a \sQuote{data.frame} of the mobility data.
}
\description{
Data contains daily mobility data between all the cities, with the columns indicating the origin and the rows indicating the destination and date.
}
\seealso{
Other datasets: 
\code{\link{coordinates}},
\code{\link{incidence}},
\code{\link{population0}},
\code{\link{population}},
\code{\link{v_by_g_day}}
}
\concept{datasets}
