% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myround.R
\name{myround}
\alias{myround}
\alias{mysignif}
\title{Round values, returning a character string keeping trailing zeros}
\usage{
myround(x, digits = 1)

mysignif(x, digits = 1)
}
\arguments{
\item{x}{a numeric vector}

\item{digits}{integer indicating the number of decimal places
('myround') or significant digits ('signif') to be used.  For 'myround',
negative values are allowed}
}
\value{
a character vector
}
\description{
Fixes the formatting issue that 'round' and 'signif' drop trailing zeros.
}
\examples{
myround(1.3001,3)
mysignif(1.3001,3)

}
\references{
'myround' is adapted from 'broman::myround'. Similar functionality
is provided by 'finalfit::round_tidy'.
}
\concept{utilities}
