% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_li.R
\name{par_li}
\alias{par_li}
\title{Parameter transform for li20v3 model}
\usage{
par_li(
  pars,
  unitOneInterval = NULL,
  unitTwoInterval = NULL,
  U,
  dir = c("from_li", "to_li")
)
}
\arguments{
\item{pars}{names parameter vector}

\item{unitOneInterval}{subset of "theta","sigma_SE","tau","E_0","Iu_0" which are unit-specific}

\item{unitTwoInterval}{subset of "alpha","Beta","mu","Z","D","Td" which are unit-specific}

\item{U}{number of units}

\item{dir}{direction, "from_li" or "to_li"}
}
\value{
transformed parameter vector
}
\description{
Transform li20 parameters to and from ranges used by Li et al
}
\examples{
par_li1 <- c(
alpha_be1 = 0.09389817,
Beta_be1  = 0.77943975,
mu_be1    = 0.96682356,
Z_be1     = 0.11650832,
D_be1     = 0.93467362,
Td_be1    = 0.80000000,
alpha_af1 = 0.46182906,
Beta_af1  = 0.13459503,
mu_af1    = 0.68640173,
Z_af1     = 0.04444878,
D_af1     = 0.12646338,
Td_af1    = 0.20000000,
theta1    = 0.89626861,
tau1      = 0.91718058,
sigma_SE1 = 0.13483710, # when dt=1/4, size_SE1=0.550023483/6, sigma_SE1 should be 0.1348371
E_01      = 0.44362243,
Iu_01     = 0.98409838
)
par_nat <- par_li(par_li1,U=5,dir="from_li")
par_li2 <- par_li(par_nat,U=5,dir="to_li")
}
