% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist.R
\name{plot_dist}
\alias{plot_dist}
\title{Mobility data plot}
\usage{
plot_dist(
  day = 1,
  city_index = 1:373,
  U = NULL,
  linewidth = 1/50000,
  arrowsize = 0.03,
  pointsize = 1,
  mob_modify_factor = 0
)
}
\arguments{
\item{day}{indicate which day will be presented on the plot}

\item{city_index}{indicate which cities will be presented, if U is not given}

\item{U}{number of the cities}

\item{linewidth}{width of the lines}

\item{arrowsize}{size of the arrows}

\item{pointsize}{size of the points}

\item{mob_modify_factor}{A multiplicative factor, which is used to control the magnitude of adjustment for the mobility data}
}
\value{
a plot of the mobility data
}
\description{
Generate a plot of the mobility data.
}
