% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_li.R
\name{plot_li}
\alias{plot_li}
\title{Plotting li20 or li23 data}
\usage{
plot_li(x, U = 373, limit = 4, order = "population")
}
\arguments{
\item{x}{a \code{spatPomp} object generated by li20(), li23(), li23v2(), etc}

\item{U}{number of the cities}

\item{limit}{the upper limit of case numbers in the plot, on a log10 scale}

\item{order}{the order of the units in the plot}
}
\value{
a \code{ggplot} plot of class \sQuote{gg} and \sQuote{ggplot} visualizing
the time series data over multiple spatial units via a tile-plot.
}
\description{
Visualize li20 or li23 COVID-19 data, with Wuhan in red.
Used to generate figures for the manuscrip and supplement of Li et al (2023).
}
