% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_li20.R
\name{plot_li20}
\alias{plot_li20}
\title{Plotting li20 data}
\usage{
plot_li20(x, U = 373, limit = 4, order = "population")
}
\arguments{
\item{x}{a \code{spatPomp} object}

\item{U}{number of the cities}

\item{limit}{the upper limit of case numbers in the plot}

\item{order}{the order of the units in the plot}
}
\value{
a \code{ggplot} plot of class \sQuote{gg} and \sQuote{ggplot} visualizing
the time series data over multiple spatial units via a tile-plot.
}
\description{
Visualize li20 data, set the higher case numbers to black and the lower ones to white, 
sort the units by population size. And set the color of Wuhan to blue, Shenyang to red.
}
