% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_percentile.R
\name{plot_percentile}
\alias{plot_percentile}
\title{Make 10, 50, 90 percentile plots for li20}
\usage{
plot_percentile(
  U = 373,
  Nsim = 100,
  order = "population",
  li20_version = "v5",
  seed = 12315,
  registerDoRNG = 3123465
)
}
\arguments{
\item{U}{number of the cities}

\item{Nsim}{number of simulations}

\item{order}{the order of the units in the plot}

\item{li20_version}{choose which version of the model will be used}

\item{seed}{the seed used for simulations}

\item{registerDoRNG}{the seed used to register the doRNG foreach backend}
}
\value{
a \code{ggplot} plot of class \sQuote{gg} and \sQuote{ggplot} visualizing
the time series data over multiple spatial units via a tile-plot.
}
\description{
This function makes spatiotemporal plots, with a 4-plot grid giving the 10, 50, 90 percentiles and a single simulation.
}
