% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{population}
\alias{population}
\title{Population}
\value{
a \sQuote{data.frame} of the population.
}
\description{
Population
}
\seealso{
Other datasets: 
\code{\link{coordinates}},
\code{\link{incidence}},
\code{\link{mobility}},
\code{\link{population0}},
\code{\link{v_by_g_day}}
}
\concept{datasets}
