% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population0.R
\name{population0}
\alias{population0}
\title{Initial population}
\value{
a \sQuote{data.frame} of the population.
}
\description{
Population at time 0. Subsequent updates due to movement are 
in 'population'
}
\references{
\li20
}
\seealso{
Other datasets: 
\code{\link{coordinates}},
\code{\link{incidence}},
\code{\link{mobility}},
\code{\link{population}},
\code{\link{v_by_g_day}}
}
\concept{datasets}
