% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R0.R
\name{R0}
\alias{R0}
\title{R0 for li20 model}
\usage{
R0(params, be = T)
}
\arguments{
\item{params}{parameter vector}

\item{be}{logical vector to calculate using parameters before Jan 23}
}
\value{
R0 value
}
\description{
Calculate  R_0 = alpha beta D + (1-alpha) mu beta D
for the model of Li et al
}
\examples{
## R0 for mle
po_mle <- li20v3(U=5,version="MLEperiod3")
p_mle <- coef(po_mle)
R0(p_mle,be=TRUE)
R0(p_mle,be=FALSE)
## R0 for li20
po_li <- li20v3(U=5,version="li20period3")
p_li <- coef(po_li)
R0(p_li,be=TRUE)
R0(p_li,be=FALSE)

}
