% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_by_g_day.R
\name{v_by_g_day}
\alias{v_by_g_day}
\title{Mobility data modification}
\value{
a \sQuote{data.frame} of the small numbers that are used to modify the mobility data.
}
\description{
Small numbers that are used to modify the mobility data. Li et al. use a multiplicative factor(theta)
to adjust the mobility data, but it won't change the zeros, and more modifications are needed.
}
\references{
\li20
}
\seealso{
Other datasets: 
\code{\link{coordinates}},
\code{\link{incidence}},
\code{\link{mobility}},
\code{\link{population0}},
\code{\link{population}}
}
\concept{datasets}
