% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_percentile.R
\name{plot_percentile}
\alias{plot_percentile}
\title{Make 10, 50, 90 percentile plots for li20 or li23}
\usage{
plot_percentile(
  U = 373,
  Nsim = 100,
  order = "population",
  model = "li23",
  seed = 12315,
  registerDoRNG = 3123465,
  params
)
}
\arguments{
\item{U}{number of the cities}

\item{Nsim}{number of simulations}

\item{order}{the order of the units in the plot}

\item{model}{choose which model will be used}

\item{seed}{the seed used for simulations}

\item{registerDoRNG}{the seed used to register the doRNG foreach backend}

\item{params}{a named numeric vector or a matrix with rownames}
}
\value{
a \code{ggplot} plot of class \sQuote{gg} and \sQuote{ggplot} visualizing
the time series data over multiple spatial units via a tile-plot.
}
\description{
This function makes spatiotemporal plots, with a 3-plot grid giving the 10, 50, 90 percentiles.
}
